+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.0]
+## [0.5.1] - 2022-08-12
+### Added
+- `attribute` method added to `Column` returning the value of the attribute, if 
+  available, or a default value (null if not provided) otherwise.
+- `label` method added to `Column` returning A human-readable label for the 
+  column, specified as the `SHORT_LABEL` or `LABEL` attribute; if these are not 
+  provided, the label is derived by expanding the name of the column.
+- Column attribute values are translated to expected type (from string).
+
+## [0.5.0] - 2022-06-16
 ### Added
 - Xls reader produces numeric values conforming to Xlsx reader.
 - Xls reader correctly estimates number of lines in sheet.